# Small sampling optimizations