I pointed out in https://github.com/ggerganov/llama.cpp/discussions/5073 that there seems to be some not completely insignificant overhead with current sampling operations:

<img width="700" alt="image" src="https://github.com/ggerganov/llama.cpp/assets/66376113/a1291ae7-2ea5-47f0-9398-4dfe40ee32fa">

With the change I made in this PR, I managed to reduce 15ms per token (for sampling over the whole initial vocabulary) to 11ms per token:

`llama_print_timings:      sample time =   11927.89 ms /  1024 runs   (   11.65 ms per token,    85.85 tokens per second)`

Efficient top k sorting is probably next.