@kalomaze 2024-01-22
![image](https://github.com/ggerganov/llama.cpp/assets/66376113/fba958ab-4f02-4321-b8fd-fad9546ba2b0)
...are my eyes deceiving me or did the prompt evaluation speed positively improve as a result of this? Shouldn't that phase be disconnected from the sample softmax function entirely? Or perhaps it's some weird CPU scheduling thing?