@kalomaze 2024-01-22
> Which change exactly results in the improvement? Most of these are noops, except the temperature check - but I don't think it can amount to that amount

Calculating the exp and sum in one pass, in the softmax function