@ggerganov 2024-01-22
I don't get it - it's the same thing before and after, except a temporary variable. It cannot have any effect on the performance