@kalomaze 2024-01-22
> I don't get it - it's the same thing before and after, except a temporary variable. It cannot have any effect on the performance

I might have tricked myself into thinking it did something because of a build flag or something. Prompt eval speed is consistent if I abandon the commits and rebuild from scratch, so it's probably that