@kalomaze 2024-01-22
Ok, this is odd.

If topk=31999:
`llama_print_timings:      sample time =    5220.53 ms /   512 runs   (   10.20 ms per token,    98.07 tokens per second)`
If topk=32000:
`llama_print_timings:      sample time =    8418.14 ms /   512 runs   (   16.44 ms per token,    60.82 tokens per second)`

1.6x faster if we sort _almost_ the whole vocabulary, but not exactly?