@kalomaze 2024-01-22
![image](https://github.com/ggerganov/llama.cpp/assets/66376113/5d35ceeb-d10c-4f00-ac12-2d88a73baee4)
With Repetition Penalty on and topk=32000 (so, the worst case scenario), ~30% of my generation time is spent on sampling for 7b q8_0.
I'm thinking that the ideal solution may be to move the sampling operations to GPU to minimize idle time