@ggerganov 2024-01-22
But what's the point of `topk = 32000` - it's a noop since all candidates remain