@kalomaze 2024-01-22
> But what's the point of `topk = 32000` - it's a noop since all candidates remain

topk=32000 is equivalent in terms of speed to topk=0 (which defaults to the full vocabulary in the same way), so something is presumably up with how the softmax function sorts / does sort checks later on?

![image](https://github.com/ggerganov/llama.cpp/assets/66376113/757d1e9e-f4ac-483e-a4f1-dc9bf40be220)