@cmp-nct 2024-01-22
I didn't notice this was already in work.
I found kalomaze's post strange as in 100 times slowness, it's the top-k sorting not the softmax.

80-90% of the large-k latency comes from sorting/partial sorting. Anything that comes after is almost irrelevant.
Check my PR, it's about 30% faster - maybe more for a slower CPU
https://github.com/ggerganov/llama.cpp/pull/5085

The real way forward imho is a dyn-k function that takes in the sampling settings and estimates a proper top-k, then the partial sort will cause a 2-3 times reduction in latency.