@kalomaze 2024-01-28
I believe this PR is made redundant by the other work done on topk/minp sorting optimizations.